warning fixes for visit_vtk directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23631 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/visit_vtk/full/vtkVolumeFromCSGVolume.C
@@ -216,7 +216,7 @@ vtkVolumeFromCSGVolume::ConstructDataSet(vtkCellData *inCD,
     for (int i = 0 ; i < nshapes ; i++)
     {
         vector<vtkCSGFixedLengthBitField> *tagList = shapeTags[i];
-        for (int j = 0; j < tagList->size(); j++)
+        for (size_t j = 0; j < tagList->size(); j++)
         {
             if (tagList->operator[](j).TestBit(VTK_CSG_MAX_BITS) == false)
                tags->push_back(tagList->operator[](j));
